requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
coverage[toml]==7.2.7
dj-database-url==2.1.0
django-extensions==3.2.3
django==3.2.21
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
exceptiongroup==1.1.3; python_version < '3.11'
flake8==5.0.4
importlib-metadata==4.2.0; python_version < '3.8'
iniconfig==2.0.0; python_version >= '3.7'
mccabe==0.7.0; python_version >= '3.6'
packaging==23.1; python_version >= '3.7'
pluggy==1.2.0; python_version >= '3.7'
pycodestyle==2.9.1; python_version >= '3.6'
pyflakes==2.5.0; python_version >= '3.6'
pyjwt==2.8.0; python_version >= '3.7'
pytest-cov==4.1.0
pytest==7.4.2; python_version >= '3.7'
pytz==2023.3.post1
sqlparse==0.4.4; python_version >= '3.5'
tomli==2.0.1
typing-extensions==4.7.1; python_version < '3.11'
zipp==3.15.0; python_version >= '3.7'
